[FSDP] [cleanup] Move state dict tests to test_fsdp_state_dict.py
